Update ACF SVG Icon plugin to version 1.1.1
This update includes bumping the version of the ACF SVG Icon plugin to 1.1.1 and uncommenting the functionality of 'acf_sprite_svg'. The function retrieves an SVG icon from a sprite and creates the necessary HTML for displaying it. Minor adjustments were also made to manage file path and cache data, enhancing overall performance.