Update ACF SVG Icon plugin to version 1.1.4
This commit includes changes to a number of PHP functions within the plugin. It primarily implements a new method for generating the _file_url by using get_theme_file_uri() in combination with the existing 'file' parameter. Moreover, a number of adjustments have been made to make the code cleaner and tighter, such as simplifying array key values and removing unnecessary comments. In addition, it modifies the JavaScript part of the plugin, mainly formatting changes and removing console logs.